weather-vsz2/requirements.txt
flask
pymongo
python-dotenv
utils
requests